refactor: :art: mini_exec.h made which have demo struct info
mini_exec.h made which have demo struct info
