+++ executer.c
@@ -6,51 +6,11 @@
 /*   By: minjacho <minjacho@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/01 15:37:17 by minjacho          #+#    #+#             */
-/*   Updated: 2024/01/01 20:28:34 by minjacho         ###   ########.fr       */
+/*   Updated: 2024/01/02 11:44:26 by minjacho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "demo_minishell.h"
-
-typedef struct	s_redirect_node
-{
-	int						type;
-	char					*file_name;
-	struct s_redirect_node	*left;
-	struct s_redirect_node	*right;
-}	t_redirect_node;
-
-typedef struct	s_simple_cmd_node
-{
-	char	*path;
-	char	**argv;
-}	t_simple_cmd_node;
-
-typedef struct	s_cmd_node
-{
-	t_redirect_node		*redirect;
-	t_simple_cmd_node	*simple_cmd;
-}	t_cmd_node;
-
-typedef struct	s_pipe_node
-{
-	t_cmd_node			*cmd;
-	struct s_pipe_node	*next_pipe;
-}	t_pipe_node;
-
-typedef struct	s_pstat
-{
-	pid_t	pid;
-	int		exit_stat;
-}	t_pstat;
-
-typedef enum	e_redirect_type
-{
-	T_INPUT,
-	T_OUTPUT,
-	T_APPEND,
-	T_HEREDOC
-}	t_redirect_type;
+#include "mini_exec.h"
 
 void	redirect_input(char *file_name, int type)
 {
